Refactor Livewire/ViewsManager to Livewire/Views/Create
